Apply PlutoUI struct style
Previously had the TTTBoard constructor be a single function that created the HTML with the event listener and js value update.  In PlutoUI, interactive elements are defined as structs with internal fields distinct from the values that update from the values saved in the bound variable.  To view and get values from the element you define Base.show and Bonds.get, etc...  This should make the behavior better for the static HTML export where previously the bound variables had missing values.